requirements.txt
accelerate==0.30.1
diffusers==0.24.0
einops==0.8.0
huggingface_hub==0.23.1
imageio==2.31.6
matplotlib==3.7.2
numpy==2.0.0
opencv_python==4.7.0.72
opencv_python_headless==4.9.0.80
pandas==2.0.2
Pillow==9.4.0
safetensors==0.4.3
skimage==0.0
spacy==3.7.1
torch==2.0.0
torchvision==0.15.0
transformers==4.41.0
en-core-web-sm==3.7.0